+++ packages/orq-rc/src/orq_ai_sdk/models/createbudgetop.py
@@ -236,5 +236,5 @@ class CreateBudgetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-10-22T18:00:57.978Z")
+    updated: Optional[datetime] = parse_datetime("2025-10-23T06:53:41.205Z")
     r"""The date and time the resource was last updated"""