+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -230,7 +230,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K86H9VZWWD0GPDY0JBKQ9RDQ"
+        "01K87XGS2CZMEE5CSDDD7WZ647"
     )
     r"""The id of the resource"""