+++ packages/orq-rc/RELEASES.md
@@ -3529,3 +3529,13 @@ Based on:
 - [python v3.14.0-rc.48] packages/orq-rc
 ### Releases
 - [PyPI v3.14.0-rc.48] https://pypi.org/project/orq-ai-sdk/3.14.0-rc.48 - packages/orq-rc
+
+## 2025-10-23 06:55:34
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.638.1 (2.729.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.14.0-rc.49] packages/orq-rc
+### Releases
+- [PyPI v3.14.0-rc.49] https://pypi.org/project/orq-ai-sdk/3.14.0-rc.49 - packages/orq-rc